+++ src/matchers.ts
@@ -17,3 +17,15 @@ const binaryRegExp = new RegExp(regexes.binary);
 export function isBinaryString(value: string): boolean {
   return binaryRegExp.test(value);
 }
+
+export function isASCIIString(value: string) {
+  for (let i = 0; i < value.length; i++) {
+    const charCode = value.charCodeAt(i);
+
+    if (charCode < 0 || charCode > 126) {
+      return false;
+    }
+  }
+
+  return true;
+}