+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added decimal to hex replace conversion
 - Added binary to hex replace conversion
+- Added ASCII to hex replace conversion
 
 ## [0.2.0]